commonize some of the 330/332 cdh shim code
