+++ sql-plugin/src/main/spark330cdh/scala/com/nvidia/spark/rapids/shims/SparkShims.scala
@@ -19,32 +19,4 @@
 spark-rapids-shim-json-lines ***/
 package com.nvidia.spark.rapids.shims
 
-import com.nvidia.spark.rapids._
-
-import org.apache.spark.sql.execution.SparkPlan
-import org.apache.spark.sql.execution.command.{CreateDataSourceTableAsSelectCommand, DataWritingCommand, RunnableCommand}
-
-object SparkShimImpl extends Spark330PlusNonDBShims with AnsiCastRuleShims {
-  override def getDataWriteCmds: Map[Class[_ <: DataWritingCommand],
-      DataWritingCommandRule[_ <: DataWritingCommand]] = {
-    Seq(GpuOverrides.dataWriteCmd[CreateDataSourceTableAsSelectCommand](
-    "Create table with select command",
-    (a, conf, p, r) => new CreateDataSourceTableAsSelectCommandMeta(a, conf, p, r))
-    ).map(r => (r.getClassFor.asSubclass(classOf[DataWritingCommand]), r)).toMap
-  }
-
-  override def getRunnableCmds: Map[Class[_ <: RunnableCommand],
-      RunnableCommandRule[_ <: RunnableCommand]] = {
-    Map.empty
-  }
-
-  override def getExecs: Map[Class[_ <: SparkPlan], ExecRule[_ <: SparkPlan]] = {
-
-    val newRule = org.apache.spark.sql.hive.cdh.HiveOverrides.getHiveTableExecRule()
-    // scala compiler gets confused if we don't give an explicit type to the class here.
-    // ... perhaps there is a cleaner way to do this
-    val ruleClass: Class[_ <: SparkPlan] = newRule.getClassFor.asSubclass(classOf[SparkPlan])
-    super.getExecs ++ Seq(ruleClass -> newRule).toMap
-  }
-
-}
+object SparkShimImpl extends Spark33cdhShims {}
